backend/requirements.txt
python-dotenv==1.0.1
langchain==0.1.8
fastapi==0.109.2
pydantic==2.6.1
starlette==0.36.3
supabase==2.3.5
openai==1.12.0
requests==2.31.0
beautifulsoup4==4.12.3
pillow==10.2.0
pydub==0.25.1
moviepy==1.0.3
pytube==15.0.0
storage3==0.7.0
tiktoken==0.6.0